// shared/achievements.d.ts
export interface Achievements {
  AssistedKills: number;
  CampsCleared: number;
  Deaths: number;
  DivineSpree: number;
  DoubleKills: number;
  FireGiantKills: number;
  FirstBloods: number;
  GodLikeSpree: number;
  GoldFuryKills: number;
  Id: number;
  ImmortalSpree: number;
  KillingSpree: number;
  MinionKills: number;
  Name: string;
  PentaKills: number;
  PhoenixKills: number;
  PlayerKills: number;
  QuadraKills: number;
  RampageSpree: number;
  ShutdownSpree: number;
  SiegeJuggernautKills: number;
  TowerKills: number;
  TripleKills: number;
  UnstoppableSpree: number;
  WildJuggernautKills: number;
}